修改权限分配功能

+++ manager/src/main/java/com/deepspc/stage/manager/system/service/impl/PermissionServiceImpl.java
@@ -3,7 +3,6 @@ package com.deepspc.stage.manager.system.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.deepspc.stage.core.enums.StageCoreEnum;
-import com.deepspc.stage.core.utils.StringUtil;
 import com.deepspc.stage.manager.common.BaseOrmService;
 import com.deepspc.stage.manager.system.entity.Menu;
 import com.deepspc.stage.manager.system.entity.Permission;
@@ -11,17 +10,14 @@ import com.deepspc.stage.manager.system.entity.UserAccess;
 import com.deepspc.stage.manager.system.mapper.MenuMapper;
 import com.deepspc.stage.manager.system.mapper.PermissionMapper;
 import com.deepspc.stage.manager.system.mapper.UserAccessMapper;
+import com.deepspc.stage.manager.system.model.AccessAssign;
 import com.deepspc.stage.manager.system.service.IPermissionService;
-import com.deepspc.stage.shiro.common.ShiroKit;
-import com.deepspc.stage.shiro.model.ShiroUser;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Map;
 
 /**
  * @author gzw
@@ -44,7 +40,6 @@ public class PermissionServiceImpl extends BaseOrmService<PermissionMapper, Perm
     @Override
     public void saveUpdatePermission(Permission permission) {
         Long permissionId = permission.getPermissionId();
-        ShiroUser shiroUser = ShiroKit.getShiroUser();
         Long relateId = permission.getRelateId();
         String permissionType = permission.getPermissionType();
         if ("01".equals(permissionType) || "02".equals(permissionType)) {
@@ -59,14 +54,8 @@ public class PermissionServiceImpl extends BaseOrmService<PermissionMapper, Perm
             }
         }
         if (null == permissionId) {
-            permission.setCreatorId(shiroUser.getUserId());
-            permission.setCreatorName(shiroUser.getUserName());
-            permission.setCreateDate(new Date());
             this.baseMapper.insert(permission);
         } else {
-            permission.setUpdatorId(shiroUser.getUserId());
-            permission.setUpdatorName(shiroUser.getUserName());
-            permission.setUpdateDate(new Date());
             this.baseMapper.updateById(permission);
         }
     }
@@ -78,20 +67,26 @@ public class PermissionServiceImpl extends BaseOrmService<PermissionMapper, Perm
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void saveUserAccess(Long selId, String assignId) {
+    public void saveUserAccess(List<AccessAssign> list) {
+        if (null == list || list.isEmpty()) {
+            return;
+        }
+        Long selId = list.get(0).getSelId();
         QueryWrapper<UserAccess> wrapper = new QueryWrapper<>();
         wrapper.eq("access_id", selId);
         userAccessMapper.delete(wrapper);
 
-        if (StringUtil.isNotBlank(assignId)) {
-            String[] ids = assignId.split(",");
-            List<UserAccess> list = new ArrayList<>();
-            for (String id : ids) {
+        List<UserAccess> userAccesses = new ArrayList<>();
+        for (AccessAssign accessAssign : list) {
+            if ("user".equals(accessAssign.getNodeType()) && null != accessAssign.getAssignId()) {
                 UserAccess ua = new UserAccess();
-                ua.setAccessId(selId);
-                ua.setUserId(Long.valueOf(id));
-                list.add(ua);
+                ua.setAccessId(accessAssign.getSelId());
+                ua.setUserId(accessAssign.getAssignId());
+                userAccesses.add(ua);
             }
+        }
+        if (!userAccesses.isEmpty()) {
+            userAccessMapper.saveBatch(userAccesses);
         }
     }
 }
